--- a/pkg/storage/store.go
+++ b/pkg/storage/store.go
@@ -701,8 +701,7 @@ type StoreConfig struct {
 // an error (which aborts all further processing for the command).
 type StoreTestingKnobs struct {
 	// TestingProposalFilter is called before proposing each command.
-	// TODO(bdarnell): Implement this when a test needs it.
-	//TestingProposalFilter storagebase.ReplicaCommandFilter
+	TestingProposalFilter storagebase.ReplicaCommandFilter
 
 	// TestingEvalFilter is called before evaluating each command. The
 	// number of times this callback is run depends on the propEvalKV