Revisiting #13869.

The leader maintains a pool of "proposal quota". Before proposing a Raft
command, we acquire proposal quota equal to the size of the
command. When all of the healthy followers have committed the entry,
quota equal to the size of the corresponding command is returned to the
pool. We maintain on Replica a map associating `storagebase.CmdIDKe`y to
the corresponding command size. Downstream of raft we extract the same
`storagebase.CmdIDKey` from the command and append the command size to
quotaReleaseQueue. This queue is subsequently used to return quota back
to the pool once all the followers have caught up.

We only consider "healthy" followers to be those that have "healthy" RPC
connections when determining if a unit of quota should be returned to the
pool.

cc @bdarnell.