@petermattis 2017-05-09
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

Can the call to `submitProposalLocked` in `refreshProposalsLocked` foul up the proposal quota adjustments? I think the answer is no because we reset the proposal quota upon Raft leadership changes. Cc @bdarnell 

---

Review status: 0 of 4 files reviewed at latest revision, 2 unresolved discussions.

---

*[pkg/storage/client_raft_test.go, line 1685 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-Kji6BHo1b6nzVRaBWt2:-Kji6BHpRIGf-bJhs8yx:bds3i29) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/client_raft_test.go#L1685)):*
> ```Go
> // slower replica catching up. By throttling write throughput we avoid having
> // to constantly catch up the slower node via snapshots. See #8659.
> func TestQuotaPool(t *testing.T) {
> ```

Nice test!

---

*[pkg/storage/client_raft_test.go, line 1696 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-Kji6G3EPEP1YMght2uP:-Kji6G3EPEP1YMght2uQ:bbeza2s) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/client_raft_test.go#L1696)):*
> ```Go
> 	mtc.replicateRange(rangeID, 1, 2)
> 
> 	mtc.stores[0].SetRaftLogQueueActive(false)
> ```

I'm guessing that you needed to disable the raft log queue because the truncations it was generating was fouling up the counting below. Might want to add a comment to that effect. Or whatever the real reason is.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-:-Kji6_KRCUCPpe65nS7o:b-g2q10b)*
<!-- Sent from Reviewable.io -->