@petermattis 2017-05-09
@tamird This is another PR that I'd be interested in seeing the impact on performance benchmarks.

---

Review status: 0 of 4 files reviewed at latest revision, 2 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-:-Kji7QrgRs7ULQ9I3uGj:b-zgulq)*
<!-- Sent from Reviewable.io -->