@bdarnell 2017-05-09
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 4 of 4 files at r1.
Review status: all files reviewed at latest revision, 3 unresolved discussions.

---

*[pkg/storage/replica.go, line 816 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-Kji7t8ZiUpLeADV6bTx:-Kji7t8_lUKNY1XYrDIB:bc220sf) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/replica.go#L816)):*
> ```Go
> 		} else {
> 			// We're becoming a follower.
> 			r.mu.proposalQuota = nil
> ```

Another thread could be blocked in `quotaPool.acquire` when we transition to follower. We need to close the old pool's channel to unblock all waiting threads. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-:-KjiA9aIj1feV65CZh54:bnfp4nl)*
<!-- Sent from Reviewable.io -->