@tamird 2017-05-09
Indeed. Good stuff, @irfansharif!

---

Reviewed 4 of 4 files at r1.
Review status: all files reviewed at latest revision, 11 unresolved discussions.

---

*[pkg/storage/client_raft_test.go, line 1696 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-Kji6G3EPEP1YMght2uP:-KjiTTvzajtHls1q-SsN:bsskfqa) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/client_raft_test.go#L1696)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

I'm guessing that you needed to disable the raft log queue because the truncations it was generating was fouling up the counting below. Might want to add a comment to that effect. Or whatever the real reason is.
</blockquote></details>

also, consider doing this by iterating through the stores:
```
for _, store := range mtc.stores {
  store.SetRaftLogQueueActive(false)
}
```

---

*[pkg/storage/client_raft_test.go, line 1707 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KjiU2wWbvLP53aR-CG_:-KjiU2wXWZxtNcij9YBh:bmmxsdk) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/client_raft_test.go#L1707)):*
> ```Go
> 	}
> 
> 	repl1, err := mtc.stores[0].GetReplica(rangeID)
> ```

This assumes that replica 1 is the leader. Could you make that more explicit? I think this test may end up flaky under stress.

You might also consider using a more descriptive variable name here, perhaps `leaderRepl`? Similarly, s/repl3/followerRepl/ below?

---

*[pkg/storage/client_raft_test.go, line 1743 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KjiUZHstnZy8vENrIA2:-KjiUZHstnZy8vENrIA3:b-faaz6) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/client_raft_test.go#L1743)):*
> ```Go
> 		defer close(ch)
> 		if _, err := client.SendWrapped(context.Background(), repl1, incArgs); err != nil {
> 			t.Fatal(err)
> ```

It is not permitted to call `t.Fatal` in the non-main goroutine. Consider changing `ch` to a `chan error` and sending the error on it, and then checking it in the main goroutine.

---

*[pkg/storage/client_raft_test.go, line 1748 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KjiUgUwpfd8pfsqOMB2:-KjiUgUwpfd8pfsqOMB3:b-bserkg) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/client_raft_test.go#L1748)):*
> ```Go
> 	}()
> 
> 	ticker := time.After(15 * time.Millisecond)
> ```

there's no need for this local, it can be inlined below

---

*[pkg/storage/client_raft_test.go, line 1751 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KjiUjs5kcSgtVlubFHd:-KjiUjs5kcSgtVlubFHe:bmpz8kq) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/client_raft_test.go#L1751)):*
> ```Go
> 	select {
> 	case <-ch:
> 		t.Fatal(errors.New("write not throttled by the quota pool"))
> ```

`Fatal` just takes `interface{}`, so you can pass a string.

---

*[pkg/storage/quota_pool.go, line 17 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KjiV2DPDYzJvtiVGN-p:-KjiV2DPDYzJvtiVGN-q:b-wn1lqv) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/quota_pool.go#L17)):*
> ```Go
> /*
>  *
>  * Copyright 2014, Google Inc.
> ```

where did this code come from?

---

*[pkg/storage/quota_pool.go, line 58 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KjiVWBkXVDYWCIlu8Ml:-KjiVWBkXVDYWCIlu8Mm:b-o50iys) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/quota_pool.go#L58)):*
> ```Go
> 	// TODO(peter): This setting needs additional thought. Should it be adjusted
> 	// dynamically?
> 	defaultProposalQuota = 1000
> ```

how come this constant is in this file, while it is only used elsewhere?

---

*[pkg/storage/quota_pool.go, line 64 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KjiVfB7Cqe0Zxx9k5yc:-KjiVfB7Cqe0Zxx9k5yd:bn2iwo6) ([raw file](https://github.com/cockroachdb/cockroach/blob/aabef51566cebad161202b0822035b1e16fb013a/pkg/storage/quota_pool.go#L64)):*
> ```Go
> 	syncutil.Mutex
> 
> 	c chan int64
> ```

Can you help me understand this structure? It looks like there's always a numeric value parked in this channel, which is read and written under lock -- why do we need a channel at all? Is it just for composition with context cancellation? The pattern is unusual enough that I think it'd benefit from a comment.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-:-KjiXiwk8kRWtRKWrisH:bhmfjxv)*
<!-- Sent from Reviewable.io -->