@irfansharif 2017-05-25
@tschottdorf Added tests + benchmarks.

---

Review status: 4 of 9 files reviewed at latest revision, 12 unresolved discussions.

---

*[pkg/storage/client_raft_test.go, line 1763 at r16](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KkWOnM34fZSlb8VK49A:-KkmRyPQMp5jQ-N8vqg5:bhuklo9) ([raw file](https://github.com/cockroachdb/cockroach/blob/4c5d9680a9e11a62e6888b61494fe155333429d1/pkg/storage/client_raft_test.go#L1763)):*
<details><summary><i>Previously, irfansharif (irfan sharif) wrote…</i></summary><blockquote>

>  You could query the quota pool though, is that possible?

Yes, didn't have a need for this earlier but re-added.

> You could query the quota pool though and loop until it's certifiably below the size of the command you propose by simply by making the key big enough to estimate a lower bound on its size.

but that doesn’t exactly work, no? If I know the command is >= 100 bytes (let’s say it’s 105 bytes) and my pool has 101 quota left - what is _supposed_ to be my _second last_ command will block indefinitely. similarly a test that would compute the size of the command (letting the first one through, calculating the difference in the quota available) and then use that to block all but one command doesn’t account for node liveness heartbeats (non-zero sized commands) and/or log truncations. I'm OK with going this way splitting a range so to not have to worry about node liveness heartbeats + log truncations like before. Is it possible to have variations in command sizes for the same command?
</blockquote></details>

Updated.

	// In order to verify write throttling we insert a value 3/4th the size of
	// total quota available in the system. This should effectively go through
	// and block the subsequent insert of the same size. We check to see whether
	// or not after this write has gone through by verifying that the total
	// quota available has decreased as expected.
	//
	// Following this we unblock the 'slow' replica allowing it to catch up to
	// the first write. This in turn releases quota back to the pool and the
	// second write, previously blocked by virtue of there not being enough
	// quota, is now free to proceed. We expect the final quota in the system
	// to be the same as what we started with.

---

*[pkg/storage/quota_pool.go, line 78 at r16](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KkT6KsbYGTOfH1vN6nA:-Kl-9Ww04CcYGCom_OVg:b-m9aw4w) ([raw file](https://github.com/cockroachdb/cockroach/blob/4c5d9680a9e11a62e6888b61494fe155333429d1/pkg/storage/quota_pool.go#L78)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

I have a theory about why the goroutine allocation is cheap (untainted by any actual research into the implementation). Goroutines live completely outside the GC heap: they can never be referenced by any GCable object (anything that could be referenced is allocated on the heap instead), so when the goroutine exits the stacks can be placed directly back on a freelist. They consume a few kilobytes while they're alive, but short-lived goroutines don't create GC pressure like other short-lived allocations. 
</blockquote></details>

Changed to using a FIFO model to avoid starvation, channels only.

      ~ go test -tags fifo -bench . -count 10
    
      BenchmarkAcquisitions-4          3000000               483 ns/op
      BenchmarkAcquisitions-4          3000000               497 ns/op
      BenchmarkAcquisitions-4          3000000               490 ns/op
      BenchmarkAcquisitions-4          3000000               502 ns/op
      BenchmarkAcquisitions-4          3000000               491 ns/op
      BenchmarkAcquisitions-4          3000000               496 ns/op
      BenchmarkAcquisitions-4          3000000               487 ns/op
      BenchmarkAcquisitions-4          3000000               505 ns/op
      BenchmarkAcquisitions-4          3000000               561 ns/op
      BenchmarkAcquisitions-4          3000000               498 ns/op
      PASS
      ok      github.com/irfansharif/qpool    20.883s

      ~ benchstat perf-cond.txt perf-fifo.txt
      name            old time/op  new time/op  delta
      Acquisitions-4   630ns ± 0%   494ns ± 2%  -21.49%  (p=0.000 n=8+9)

---

*[pkg/storage/quota_pool.go, line 109 at r18](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KklrrFwAxvdRqwDCR1k:-Kkvar5ihWZWmC9mW7pa:b-xsmg0w) ([raw file](https://github.com/cockroachdb/cockroach/blob/260171425203cda76c51976f96c5345d3f68b2f4/pkg/storage/quota_pool.go#L109)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary><blockquote>

I'm concerned about starvation for large commands. If there's a constant flow of small commands they may grab all the quota without ever allowing the available quota to grow to the level necessary for a large command to proceed. I think we want more of a FIFO model in which commands can claim their position in line and then block the queue until enough quota is available for them. That probably simplifies the implementation too. 
</blockquote></details>

Using a FIFO model now, PTAL.

---

*[pkg/storage/replica.go, line 90 at r16](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KkTBbRjIaPA0XfzeMM-:-Kl-9ru3fBWWjdLgf5fx:bgjp008) ([raw file](https://github.com/cockroachdb/cockroach/blob/4c5d9680a9e11a62e6888b61494fe155333429d1/pkg/storage/replica.go#L90)):*
<details><summary><i>Previously, tschottdorf (Tobias Schottdorf) wrote…</i></summary><blockquote>

ah, makes sense. So we want `max_size <= pool_size < raftLogMaxSize` (where it's not clear that 1MB of raft log size corresponds to roughly 1MB of pool size. It should be slightly less).
</blockquote></details>

set `pool_size` to be `raftLogMaxSize/4` and for commands that are larger than this amount we allow it to go through if it's able to acquire quota equal to the max proposal quota. The current implementation can be easily extended to accommodate dynamic changing of total quota but the current fixed size should suffice.

---

*[pkg/storage/replica.go, line 867 at r18](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-KklcFJEVOKdWvJ_0Lln:-KkmSbnhzlRmAcKIXZnf:bsybv8v) ([raw file](https://github.com/cockroachdb/cockroach/blob/260171425203cda76c51976f96c5345d3f68b2f4/pkg/storage/replica.go#L867)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

A `nil` slice is usually equivalent to a zero length slice. I'm wondering why this is necessary.
</blockquote></details>

It is not, removed.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/15802#-:-Kkvr4rFq17jHWFOmFrT:bi5wjtf)*
<!-- Sent from Reviewable.io -->